🚚 move: 커스텀 유효성 검증 common 패키지 하위로 이동
